Added rudimentary pygame support and example
The naive implementation of render_grid will recalculate cell sizes
and padding every time it is called.  A future version of the function
should pre-calculate the widths and padding and return a function that
simply calls the render_cell function with the appropriate sizes.

In order to use this you will have to install the 'pygame'
extras_requires with this package as Horton is still useful for many
applications without it.