Added basic controls to the demo
You can now adjust the grid size and width in the demo using the '+'
and '-' keys to adjust padding and the arrow keys to adjust the size.
It's kind of fun to play around with and my rendering calculations do
some fun things when you get close to 1 on the various axis.